cpu: Physical register structural + flat indexing
Mimic the changes done on the architectural register indexes
on the physical register indexes. This is specific to the O3
model.
The structure, called PhysRegId,  contains a register class,
a register index and a flat register index. The flat
register index is kept because it is useful in some cases
where the type of register is not important (dependency
graph and scoreboard for example).
Instead of directly using the structure, most of the code
is working with a const PhysRegId* (typedef to PhysRegIdPtr).
The actual PhysRegId objects are stored in the regFile.

Change-Id: Ic879a3cc608aa2f34e2168280faac1846de77667
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>