cpu: Added interface for vector reg file
This patch adds some more functionality to the cpu model and the arch to
interface with the vector register file.

This change consist mainly in augmenting ThreadContexts and
ExecContexts with calls to get/set full vectors, underlying
microarchitectural elements  or lanes. Those are meant to interface with
the vector register file. All classes that implement this interface also
get an appropriate implementation.

This requires implementing the vector register file for the different
models using the VecRegContainer class.

This change set also updates the Result abstraction to contemplate the
possibility of having a vector as result.

The changes also affect how the remote_gdb connection works.

There are some (nasty) side effects, as the need to define dummy values
for architectures that do not implement vector extensions.
Initialisation of numPysVecRegs in the test configurations,

Nathanael Premillieu's work with an increasing number of fixes and
improvements of mine.

Change-Id: Iee65f4e8b03abfe1e94e6940a51b68d0977fd5bb
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>