base: Allocators definition
Define a basic allocator. The whole point of the allocator is to be
able to have std::vector of objects that contain stats and construct
them passing parametres to the constructor.

This patch defines a basic allocator that receives a variable amount
of parametres, and when used to construct the target object, it passes
the parametres to the constructor.

The immediate goal of this patch is having the o3 LSQUnit constructor
have the sizes of the LQ and SQ as parametres, and be able to define
the threads in the o3 LSQ as a std::vector LSQUnit that is constructed
passing the appropriate value to the parametres.

Change-Id: Id57184707dedc6d54092f8a39fb82493c46dc804
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>