arch-arm: Fixes to fplibFPToFixed for converting from 16b float
When converting from a 32-bit or 64-bit float we deliberately did not
recognise infinities because an infinity interpreted as a normalised
number would in any case give the right result when saturated. This
is not true when converting from a 16-bit float to a 32-bit or 64-bit
integer: in this case we must recognise infinities.

Also remove redundant cast from uint32_t fplibFPToFixed(uint16_t, ...).

Change-Id: Ib6d26760ea9a53ca4ad0a76a0b02123a2669d08b
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>